fix: make sure selenium driver isn't loaded on every test
